+++ packages/vsce/src/constants/Cics.defaults.ts
