+++ packages/vsce/src/trees/CICSTree.ts
@@ -26,7 +26,7 @@ import {
   window,
 } from "vscode";
 import { PersistentStorage } from "../utils/PersistentStorage";
-import constants from "../utils/constants";
+import constants from "../constants/Cics.defaults";
 import { getErrorCode } from "../utils/errorUtils";
 import { FilterDescriptor } from "../utils/filterUtils";
 import { InfoLoaded, ProfileManagement } from "../utils/profileManagement";
